Launch process in terminal emulator if required
If a desktop entry requires a terminal, but the current process does not
run in one, the process would previously run invisibly in the
background. Now we instead start the process in a newly launched
terminal in that case. This is especially useful when (for example)
opening text files downloaded in a browser using the system open dialog.

For now the terminal emulator to be used is specified in the
configuration file. When/if there is a standard way to select the
terminal emulator, it is conceivable to use that method instead or if no
config item is present (see
https://gitlab.freedesktop.org/xdg/xdg-utils/-/issues/84).